Add the ability to skip SAF check for privileged callers of a cross memory server. (#91)
The purpose of this enhancement is to allow calling cross-memory
servers from system exits.

A new public function has been added that allows passing flags to
cross-memory calls. At the moment there is only one flag - the "skip SAF
check" flag. The flags makes it possible to skip the SAF check if the
caller is either in:
* Supervisor state
* System key (key 0-7)

If the caller is in neither of those states, the flag is ignored. If needed, the users of the cross-memory server can enforce the SAF check even if the caller is privileged.

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>