Control LPA dev mode with a run-time option.
This commit adds a new control flag. The flag allows users to enable the LPA
dev mode without recompiling the module. In the LPA dev mode, the
ZWESIS01 module will always be refreshed at start up and removed from
LPA upon termination. This latter should make the memory footprint smaller in
a development environment.

If the new run-time mode is enabled, a warning message will be issued to
the SYSLOG and SYSPRINT.

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>