Verify that ZWESIS01 comes from private storage.
If the loadmodule has been copied incorrectly to STEPLIB, the z/OS loader
will fail to load it. If there is an available copy in LPA it will be
used instead. The problem with LPA is that the module may not be of the
desired version, and any IDENTIFY calls on an LPA module will cause
serious issues. This commit ensures that ZWESIS01 comes from private
storage.

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>