Module registry initial implementation
To address potential LPA shortages when multiple ZIS instances are
running on the same LPAR, a new feature--the module registry--has been
introduced.

The module registry is a set of data structures in common storage in
which eligible modules are traced. An eligible module is a module marked
with a special macro; the mark and the module names are then used to
uniquely identify the module. Once ZIS and its plug-in modules are
marked, they can be added to the registry and shared: if another ZIS
instance needs to load an identical module to the LPA, it will reuse
a previously registered (and loaded to the LPA) module instead. Sharing
identical modules should dramatically decrease the LPA footprint.

This feature must not be used when an application wants to remove its
modules from the LPA (e.g., ZIS does that in dev mode) because it will
affect any other applications using the same modules.

Fixes: #405

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>