Fix an incorrect check in the recovery router
The recovery facility might erroneously free the state cell-pool if an
ABEND occurs right after the ESTAEX has been set but before any recovery
states with the retry option have been pushed; this issued is caused
by an incorrectly coded check in the recovery router code (the check is
inverted).

The commit fixes the incorrect instruction and adds some code to leave
"bread crumbs" when the state cell-pool is deleted.

Fixes: #446
Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>