Make sure CEE3ERP is invoked in LE 31-bit XPLINK; add tests
When an ABEND occurs and there is a user-defined ESTAEX in
an LE application, the language environment must be notified
via a call to CEE3ERP; that way LE has a chance to handle
things like hitting a stack guard page. If we don't call
CEE3ERP, things can go terribly wrong.

At some point, the ZSS 31-bit build was changed to use
XPLINK and the CEE3ERP call in the recovery facility was
erroneously limited to non-XPLINK 31-bit LE environments.

This commit changes the code to call the CEE3ERP routine
in XPLINK 31-bit LE applications.

Fixes:
* zowe/zss#600
* zowe/zss#736

Signed-off-by: Irek Fakhrutdinov <ifakhrutdinov@rocketsoftware.com>